[REFACTOR] plan, room slice 분리
